metrics: move package to services
Where it belongs.
